# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ticker Lab API
  description: |
    Financial data dashboard API. Ingests ECB exchange rates daily via Frankfurter
    and serves them through a REST interface. Part of the Ticker Lab experiment
    for aggregating public financial data.
  version: 0.5.0
  contact:
    name: Ticker Lab
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Local development

tags:
  - name: system
    description: Health and operational endpoints
  - name: exchange-rates
    description: ECB exchange rate data via Frankfurter

paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      description: Returns service health status. Used by load balancers and monitoring.
      tags:
        - system
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
                timestamp: "2026-04-19T12:00:00.000Z"

  /ready:
    get:
      operationId: getReadiness
      summary: Readiness check
      description: Returns service readiness status including database connectivity. Used by orchestrators to determine if the service can accept traffic.
      tags:
        - system
      responses:
        "200":
          description: Service is ready to accept traffic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
              example:
                status: ready
                checks:
                  database: ok
                timestamp: "2026-04-19T12:00:00.000Z"
        "503":
          description: Service is not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
              example:
                status: not_ready
                checks:
                  database: error
                timestamp: "2026-04-19T12:00:00.000Z"

  /api/v1/exchange-rates/latest:
    get:
      operationId: getLatestExchangeRates
      summary: Get latest exchange rates
      description: Returns the most recent exchange rates for a given base currency from the ECB.
      tags:
        - exchange-rates
      parameters:
        - $ref: "#/components/parameters/BaseCurrency"
      responses:
        "200":
          description: Latest exchange rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRatesResponse"
              example:
                base: EUR
                date: "2026-04-17"
                rates:
                  - currency: USD
                    rate: 1.1358
                  - currency: GBP
                    rate: 0.8561
        "404":
          description: No rates found for the given base currency
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  /api/v1/exchange-rates/{date}:
    get:
      operationId: getExchangeRatesByDate
      summary: Get exchange rates for a specific date
      description: Returns exchange rates for a given base currency on a specific date from the ECB.
      tags:
        - exchange-rates
      parameters:
        - $ref: "#/components/parameters/BaseCurrency"
        - name: date
          in: path
          required: true
          description: Date in ISO 8601 format (YYYY-MM-DD). Must be a business day.
          schema:
            type: string
            format: date
          example: "2026-04-17"
      responses:
        "200":
          description: Exchange rates for the requested date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRatesResponse"
        "404":
          description: No rates found for the given date and base currency
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  /api/v1/exchange-rates/history:
    get:
      operationId: getExchangeRateHistory
      summary: Get historical exchange rates for a currency pair
      description: Returns a time series of exchange rates for a specific currency pair over a date range. Used for charts and trend analysis.
      tags:
        - exchange-rates
      parameters:
        - $ref: "#/components/parameters/BaseCurrency"
        - name: quote
          in: query
          required: true
          description: ISO 4217 quote currency code
          schema:
            type: string
            minLength: 3
            maxLength: 3
          example: USD
        - name: from
          in: query
          required: false
          description: Start date (YYYY-MM-DD). Defaults to 30 days ago.
          schema:
            type: string
            format: date
          example: "2026-01-01"
        - name: to
          in: query
          required: false
          description: End date (YYYY-MM-DD). Defaults to today.
          schema:
            type: string
            format: date
          example: "2026-04-17"
      responses:
        "200":
          description: Historical exchange rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
              example:
                base: EUR
                quote: USD
                from: "2026-01-01"
                to: "2026-04-17"
                rates:
                  - date: "2026-01-02"
                    rate: 1.0358
                  - date: "2026-01-03"
                    rate: 1.0412

components:
  parameters:
    BaseCurrency:
      name: base
      in: query
      required: false
      description: ISO 4217 base currency code. Defaults to EUR.
      schema:
        type: string
        default: EUR
        minLength: 3
        maxLength: 3
      example: EUR

  schemas:
    HealthResponse:
      title: HealthResponse
      description: Service health status
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          description: Health status indicator
          enum:
            - ok
          example: ok
        timestamp:
          type: string
          format: date-time
          description: Current server time in ISO 8601 format
          example: "2026-04-19T12:00:00.000Z"

    HistoryResponse:
      title: HistoryResponse
      description: Time series of exchange rates for a currency pair
      type: object
      required:
        - base
        - quote
        - from
        - to
        - rates
      properties:
        base:
          type: string
          description: ISO 4217 base currency code
          example: EUR
        quote:
          type: string
          description: ISO 4217 quote currency code
          example: USD
        from:
          type: string
          format: date
          description: Start date of the range
          example: "2026-01-01"
        to:
          type: string
          format: date
          description: End date of the range
          example: "2026-04-17"
        rates:
          type: array
          description: Time series of rates ordered by date ascending
          items:
            $ref: "#/components/schemas/HistoryPoint"

    HistoryPoint:
      title: HistoryPoint
      description: A single data point in a rate time series
      type: object
      required:
        - date
        - rate
      properties:
        date:
          type: string
          format: date
          description: Date of the rate
          example: "2026-01-02"
        rate:
          type: number
          description: Exchange rate on this date
          example: 1.0358

    ReadinessResponse:
      title: ReadinessResponse
      description: Service readiness status with dependency checks
      type: object
      required:
        - status
        - checks
        - timestamp
      properties:
        status:
          type: string
          description: Overall readiness status
          enum:
            - ready
            - not_ready
          example: ready
        checks:
          type: object
          description: Individual dependency check results
          required:
            - database
          properties:
            database:
              type: string
              description: Database connectivity status
              enum:
                - ok
                - error
              example: ok
        timestamp:
          type: string
          format: date-time
          description: Current server time in ISO 8601 format
          example: "2026-04-19T12:00:00.000Z"

    ExchangeRatesResponse:
      title: ExchangeRatesResponse
      description: A set of exchange rates for a base currency on a given date
      type: object
      required:
        - base
        - date
        - rates
      properties:
        base:
          type: string
          description: ISO 4217 base currency code
          example: EUR
        date:
          type: string
          format: date
          description: Date of the exchange rates (YYYY-MM-DD)
          example: "2026-04-17"
        rates:
          type: array
          description: List of exchange rates relative to the base currency
          items:
            $ref: "#/components/schemas/Rate"

    Rate:
      title: Rate
      description: A single exchange rate for a currency pair
      type: object
      required:
        - currency
        - rate
      properties:
        currency:
          type: string
          description: ISO 4217 quote currency code
          example: USD
        rate:
          type: number
          description: Exchange rate relative to the base currency
          example: 1.1358

    ProblemDetails:
      title: ProblemDetails
      description: Error response following RFC 9457
      type: object
      required:
        - type
        - title
        - status
        - detail
        - code
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: "https://tickerlab.dev/problems/not-found"
        title:
          type: string
          description: Short human-readable summary of the problem
          example: Not Found
        status:
          type: integer
          description: HTTP status code
          example: 404
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: No exchange rates found for EUR on 2026-04-20
        code:
          type: string
          description: Machine-readable error code in SCREAMING_SNAKE_CASE
          example: RATES_NOT_FOUND